Add files via upload
Dazu Huike Token is built on Ethereum infrastructure (ERC-20). Bridges to BSC, FTM and AVAX networks are also being developed in addition to the Holy Grail of Dexs, which will change the way investors view decentralized exchanges. All this made possible by our Community Owned Decentralized Token. Is a medium for traders, crypto enthusiasts, and ordinary people in general to be able to transact as currency, payment and monetary system as whole. More info on https://dazutoken.com/